// frontend/src/api/generated/models/saveRecipeRequest.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Hestia API
 * Internal API for Hestia household inventory management
 * OpenAPI spec version: 1.0.0
 */
import type { SaveRecipeRequestIngredients } from './saveRecipeRequestIngredients.ts';

export interface SaveRecipeRequest {
  /** @maxLength 255 */
  name: string;
  /** @minItems 1 */
  ingredients?: SaveRecipeRequestIngredients;
  /** @nullable */
  instructions?: string | null;
  /**
     * @maxLength 1024
     * @nullable
     */
  source_url?: string | null;
}